Add default understanding of the main method being not in the project root.
It's hard to prescribe what to do to scale a project build config that has $n$ commands, but I still find it's reliably a good idea to start with the main method in this pattern of package layout right off the bat.

You can remove the extra strings if you disagree or it's not a good fit for your project.