Create the self-symlink on 'init'.
Mostly a no-op.  But if we're being honest about how humans need to use computers: new projects start somewhere.  For me, it's copying this script.  So, yeah, if there are setup actions it can sprout that are zero-overhead?  In.

Within reason.  Eventually there will a battle between The Forces of the Kitchen Sink and the Angelic Army of Brevity, with the Plugin Revolution taking up an independent run on the throne which must be resisted at all costs.  But these are two lines that definitely save me an absurd amount of time.