Add test timeouts, overridable by env, and low by default.
The golang default of 10m is... very forgiving.  The most common way to interact with it is realize that you're careening towards it due to an obvious bug and it's either time to take a stroll to get coffee or start forming your SIGQUIT'ing kill commands.  Let's just be less smotheringly patient by default.