You know what?  Race detector: on by default.
If you don't think you need the race detector, you're probably wrong; or, you're right for about five more seconds of your development history and then you'll *become* wrong.

And adding the race detector in to a project that's too far along can be nothing short of a nightmare of embarassments and a horror march trying to backpedal to a happier place.

So.  Start safe.