Revert "You know what?  Race detector: on by default."  The costs of the race detector are high enough to be worthy of comment.
In particular, the race detector seems to raise the time needed to run a binary to a minimum of around 1 full wall-clock second.  That means tests of each package will take at least that full second.  For projects with many small packages, this becomes somewhat unpleasant.

Race detection is still incredibly useful, and important to run early and often, but having it on by default is just too expensive.

(I'm also moved by the bug in the 1.8 go release which makes builds with the race detector error out-of-box if the go install directory is not writable, to be honest.)

Signed-off-by: Eric Myhre <hash@exultant.us>